+++ src/sorting.jl
@@ -7,7 +7,7 @@ export mysortperm,mysortperm!
 
 function my_sortperm_you_should_regularily_check_if_sortperm_in_base_has_become_more_efficient!(ii,A)
     const n = length(A)
-   #ii = Array(Int64,n)
+   #ii = Array(Int,n)
 
    for i = 1:n
     ii[i] = i
@@ -23,7 +23,7 @@ end # function mysortperm!
 
 function mysortperm(A)
     const n = length(A)
-   ii = Array(Int64,n)
+   ii = Array(Int,n)
 
    for i = 1:n
     ii[i] = i
